Work with Multihashes directly (#21)
This updates the filestore to work with raw multihashes instead of CIDs, aligning with the latest versions of go-ipfs-ds-help. Every block in the filestore should already be a raw multihash, however, since they were added as merkledag.RawNode, which is a Cidv1 of type Raw, and maps nicely to a raw multihash.

Interfaces have been kept without change, but may be updated in the future to expose only Multihashes.